test(io): refactor save tests
- Remove redundant method overrides
- Rename files for consistency
